requirements.txt
Flask==1.1.2
line-bot-sdk==1.19.0
pymongo==3.11.4
dnspython==2.1.0
paho-mqtt==1.5.1
